+++ webapp/src/main/java/ar/edu/itba/paw/webapp/controller/VoteController.java
@@ -4,9 +4,12 @@ import ar.edu.itba.paw.interfaces.services.VoteService;
 import ar.edu.itba.paw.model.User;
 import ar.edu.itba.paw.model.Vote;
 import ar.edu.itba.paw.model.enums.DebateVote;
+import ar.edu.itba.paw.model.exceptions.ForbiddenVoteException;
 import ar.edu.itba.paw.webapp.dto.VoteDto;
 import ar.edu.itba.paw.webapp.form.VoteForm;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 import javax.validation.Valid;
@@ -34,6 +37,12 @@ public class VoteController {
             @Valid @NotNull(message = "missing body") final VoteForm form
     ) throws UnsupportedEncodingException {
         final String username = URLDecoder.decode(url, User.ENCODING);
+
+        final Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (!username.equals(auth.getName())) {
+            throw new ForbiddenVoteException();
+        }
+
         voteService.addVote(debateId, username, DebateVote.valueOf(form.getVote().toUpperCase()));
 
         return Response.created(null).build();
@@ -44,6 +53,11 @@ public class VoteController {
             @Valid @NotNull(message = "user param must be included") @QueryParam("user") final String url
     ) throws UnsupportedEncodingException {
         final String username = URLDecoder.decode(url, User.ENCODING);
+
+        final Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (!username.equals(auth.getName())) {
+            throw new ForbiddenVoteException();
+        }
 
         if (voteService.removeVote(debateId, username)) {
             return Response.noContent().build();
@@ -57,6 +71,11 @@ public class VoteController {
             @Valid @NotNull(message = "user param must be included") @QueryParam("user") final String url
     ) throws UnsupportedEncodingException {
         final String username = URLDecoder.decode(url, User.ENCODING);
+
+        final Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (!username.equals(auth.getName())) {
+            throw new ForbiddenVoteException();
+        }
 
         Optional<Vote> vote = voteService.getVote(debateId, username);
         if (vote.isPresent()) {
